# deploy/manifests/swagger/bookly.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Client API | Hotel API",
    "description": "A merged version hotel-endpoints.raml and client-endpoints.raml from specification. endpoints.raml were omitted, since its basically a copy of hotel-endpoints.raml. Please note that specification might not been converted 100% correctly to OAS 2.0 and there might be errors or misinterpretations. Do not hesitate to contact me if you think there is error in this documentation. Tokens for logging: {\"ID\": 99999998, \"CreatedAt\":\"2020-01-01\"}\n{\"ID\": 99999999, \"CreatedAt\":\"2020-01-01\"}",
    "version": "1.1 - 14.04.2021"
  },
  "servers": [
    {
      "url": "http://b.cebb8e0ee2ba4e9785f9.westeurope.aksapp.io/"
    },
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/api-hotel/reservations": {
      "get": {
        "security": [
          {
            "customTokenSecurityHotel": []
          }
        ],
        "description": "fetches current (and future) reservations made by clients and information regarding these clients",
        "operationId": "GET_reservations",
        "parameters": [
          {
            "name": "currentOnly",
            "in": "query",
            "description": "get reservations that are currently underway or all reservations (including ones that begin in the future)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "roomID",
            "in": "query",
            "description": "get reservations connected with room with ID equal to roomID parameter",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Number of page to show results",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of show results",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of objects containing reservation and client information related to a hotel room reservation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/reservationObject"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Room with ID equal to roomID does not belong to the hotel",
            "content": {}
          },
          "404": {
            "description": "An error containing message describing the type of error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "error": "Room with ID equal to roomID parameter does not exist"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api-client/client/reservations": {
      "get": {
        "description": "Get list of all reservations made by the client",
        "operationId": "GET_client-reservations",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Number of page to show results",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of show results",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reservation information and offerID relating to the reservation and preview information related to the hotel",
            "content": {}
          }
        },
        "security": [
          {
            "customTokenSecurityClient": []
          }
        ]
      }
    },
    "/api-client/client/reservations/{reservationID}": {
      "delete": {
        "description": "Cancel exisiting reservation provided that it hasn't started",
        "operationId": "DELETE_client-reservations-reservationID",
        "parameters": [
          {
            "name": "reservationID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {}
          },
          "400": {
            "description": "Reservations cannot be deleted unless they haven't begun",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "error": "Reservation is currently underway or already completed"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Cannot delete reservation becuase of no ownership of a reservation with ID equal to reservationID path parameter",
            "content": {}
          },
          "404": {
            "description": "Reservation with ID equal to reservationID path parameter doesn't exist",
            "content": {}
          }
        },
        "security": [
          {
            "customTokenSecurityClient": []
          }
        ]
      }
    },
...
    "/api-client/hotels/{hotelID}/offers/{offerID}/reservations": {
      "post": {
        "description": "Create new reservation",
        "operationId": "POST_hotels-hotelID-offers-offerID-reservations",
        "parameters": [
          {
            "name": "hotelID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offerID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "from",
                  "numberOfAdults",
                  "numberOfChildren",
                  "to"
                ],
                "type": "object",
                "properties": {
                  "from": {
                    "type": "string",
                    "format": "date"
                  },
                  "to": {
                    "type": "string",
                    "format": "date"
                  },
                  "numberOfChildren": {
                    "type": "integer"
                  },
                  "numberOfAdults": {
                    "type": "integer"
                  }
                },
                "example": {
                  "from": "2021-10-02T15:00:00.000Z",
                  "to": "2021-12-02T15:00:00.000Z",
                  "numberOfChildren": 1,
                  "numberOfAdults": 2
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {}
          },
          "400": {
            "description": "Cannot create a reservation if the offer is inactive or deleted or is unavailable during chosen time interval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "error": "Offer is not available - please refresh information related to the offer availability"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {}
          }
        },
        "security": [
          {
            "customTokenSecurityClient": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "reservationObject": {
        "required": [
          "client",
          "reservation",
          "room"
        ],
        "type": "object",
        "properties": {
          "reservation": {
            "required": [
              "adultsCount",
              "childrenCount",
              "fromTime",
              "offerID",
              "reservationID",
              "toTime"
            ],
            "type": "object",
            "properties": {
              "reservationID": {
                "type": "integer"
              },
              "offerID": {
                "type": "integer"
              },
              "fromTime": {
                "type": "string",
                "format": "date-time"
              },
              "toTime": {
                "type": "string",
                "format": "date-time"
              },
              "childrenCount": {
                "type": "integer"
              },
              "adultsCount": {
                "type": "integer"
              }
            }
          },
          "client": {
            "required": [
              "clientID",
              "name",
              "surname"
            ],
            "type": "object",
            "properties": {
              "clientID": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "surname": {
                "type": "string"
              }
            }
          },
          "room": {
            "required": [
              "hotelRoomNumber",
              "roomID"
            ],
            "type": "object",
            "properties": {
              "roomID": {
                "type": "integer"
              },
              "hotelRoomNumber": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "customTokenSecurityHotel": {
        "type": "apiKey",
        "description": "Custom security token given by the server administrator.",
        "name": "x-hotel-token",
        "in": "header"
      },
      "customTokenSecurityClient": {
        "type": "apiKey",
        "description": "Custom security token generated by the server. Distributed to clients upon successful login and used for requests requiring client authentication.",
        "name": "x-session-token",
        "in": "header"
      }
    }
  }
}